requirements.txt
requests
beautifulsoup4
lxml
pandas
urllib3
selenium
tqdm
python-dateutil
pyyaml
